uhd: use preprocessor directives for UHD version control
UHD 003.004.000 is not an official release, though the
API changes are getting a notable amount of use. Ifdef new
interface changes for now and re-evaluate making 003.004.000
a minimum requirement at a later time.

This also allows use of the streamer interface in the main
branch, which addresses various reports of UHD failing to
send packets at start with the old, deprecated API.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>