uhd: handle b100 / e100 overflow behaviour
B100 and E100 will react differently to overruns then the USRP2 or
N2xx devices. To handle such an overrun, we need to expect a short
packet with no error code followed by zero length packet with the
appropriate error code set. On USRP2, the event is contained
within a single packet.

To handle, if a short packet contains no error code, pull another
packet and check again for error code. If still no error code,
print error on "short packet" and continue. If there is an error
code, as expected, proceed normally and handle the error accord-
ingly.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>